implement new canvas in app (it's broken)
