+++ src/ThemeGenerator/components/Canvas/Canvas.scss
@@ -1,10 +1,14 @@
 .Canvas {
+  display: block;
   position: relative;
 
   &__canvas {
-    display: block;
     position: absolute;
-    width: 100%;
-    height: 100%;
+    left: 0;
+    top: 0;
+  }
+
+  &__tester {
+    padding: 1rem;
   }
 }
