+++ src/ThemeGenerator/components/Canvas/web-workers/mask.worker.ts
@@ -1,16 +1,21 @@
 import { DEFAULT_CANVAS_SIZE } from 'ThemeGenerator/constants'
-import { WorkerContext } from './types'
-import { createCanvas, getInitialState } from './utils'
+import { WorkerContext } from '../types'
+import { createCanvas, getInitialState, resize } from './utils'
 import { renderMask } from './render-mask'
 
 declare const self: WorkerContext
 let state = getInitialState(DEFAULT_CANVAS_SIZE)
-const cache: Record<number, ImageBitmap> = {}
+let cache: Record<number, ImageBitmap> = {}
 const { canvas, canvasContext } = createCanvas(state.size)
 
 self.onmessage = ({ data }) => {
-  const { hue, size } = data
-  state = { ...state, hue, size }
+  if (data.size !== state.size) {
+    resize(canvas, data.size)
+    cache = {}
+  }
+
+  state.hue = data.hue
+  state.size = data.size
 
   if (!state.hasRenderPending) {
     state.hasRenderPending = true
@@ -18,7 +23,11 @@ self.onmessage = ({ data }) => {
       if (!cache[state.hue])
         cache[state.hue] = renderMask(state, canvas, canvasContext)
 
-      self.postMessage({ hue: state.hue, bitmap: cache[state.hue] })
+      self.postMessage({
+        hue: state.hue,
+        size: state.size,
+        bitmap: cache[state.hue],
+      })
       state.hasRenderPending = false
     })
   }
