+++ src/ThemeGenerator/components/CanvasOld/CanvasOld.scss
@@ -1,0 +1,10 @@
+.CanvasOld {
+  position: relative;
+
+  &__canvas {
+    display: block;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+  }
+}
