+++ src/ThemeGenerator/components/CanvasOld/worker-loader.d.ts
