+++ src/ThemeGenerator/components/CanvasOld/worker.ts
