--- src/ThemeGenerator/components/CanvasTest/CanvasTest.scss
@@ -1,4 +1,0 @@
-.CanvasTest {
-  padding: 1rem;
-  display: block;
-}
