--- src/ThemeGenerator/components/CanvasTest/index.ts
@@ -1,1 +1,0 @@
-export * from './CanvasTest'
